src/test/java/edu/hw1/Task8Test.java
package edu.hw1;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class Task8Test {
    Task8 task8 = new Task8();

    @Test
    @DisplayName("Кони не бьют друг друга")
    void knightNotCaptured() {
        //given
        int[][] board = {
            {1, 0, 1, 0, 1, 0, 1, 0},
            {0, 1, 0, 1, 0, 1, 0, 1},
            {1, 0, 1, 0, 1, 0, 1, 0},
            {0, 1, 0, 1, 0, 1, 0, 1},
            {1, 0, 1, 0, 1, 0, 1, 0},
            {0, 1, 0, 1, 0, 1, 0, 1},
            {1, 0, 1, 0, 1, 0, 1, 0},
            {0, 1, 0, 1, 0, 1, 0, 1}};
        //when
        boolean isKnightCaptured = task8.knightBoardCapture(board);
        //then
        Assertions.assertThat(isKnightCaptured).isEqualTo(true);
    }

    @Test
    @DisplayName("Кони бьют друг друга")
    void knightCapture() {
        //given
        int[][] board = {
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 1, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 1},};
        //when
        boolean isKnightCaptured = task8.knightBoardCapture(board);
        //then
        Assertions.assertThat(isKnightCaptured).isEqualTo(false);
    }
}
